Beginning to organize and refactor methods into classes.
