+++ lib/date_generator.rb
@@ -1,0 +1,8 @@
+module DateGenerator
+
+    def date_generator
+      date = Date.today
+      string = date.strftime("%m%d%y")
+    end
+
+end
