+++ lib/enigma_helper_methods.rb
@@ -1,0 +1,5 @@
+require 'date'
+module EnigmaHelperMethods
+
+
+end
