+++ test/date_generator_test.rb
@@ -1,0 +1,14 @@
+require './test/test_helper'
+require 'minitest/autorun'
+require 'minitest/pride'
+
+class DateGeneratorTest < Minitest::Test
+  def setup
+    @enigma = Enigma.new
+  end
+
+  def test_date_generator_returns_todays_date
+    assert_equal "022419", @enigma.date_generator
+  end
+
+end
