+++ test/test_helper.rb
@@ -4,3 +4,4 @@ require 'minitest/autorun'
 require 'minitest/pride'
 require 'pry'
 require './lib/enigma'
+require './lib/date_generator'
